Add "label_area" attribute to classification.geojson
On top of the "label" attribute indicating the presence or absence of a given class in the classification.geojson output, a new attribute called "label_area" was integrated which indicates the surface area of the present classes in a given tile. This is particularly useful as an indicator of class abundance.